rework album page to best match old behavior
